Simplify error handling
The previous behavior was sophisticated but buggy: (1) For many errors, the line/column number was not detected properly (perhaps yapf's output format has changed over the months?) (2) The line/column/character number appears to have never worked in the presence of selections.
I have therefore replaced it by autoyapf's dumb error reporting which has proved rather robust for the past year or so.
It might be desirable to reimplement the old features in a more robust fashion, but this will require some more work.