Fixed bug in newline handling
Yapf will always add a trailing newline, even if the original code did not end on a newline. While this is good practice for source files, it may happen that we ask yapf to format a selection that does not span an entire line, and so we have to worry about this case.